moved requestSubscriptions into services folder
